created EnterMeal as a component
