+++ pages/search.js
@@ -6,7 +6,7 @@ import Recipe from "../utils/Recipe";
 import Head from "next/head";
 import Button from "react-bootstrap/Button";
 import config from "../utils/config";
-import EnterMeal from "./form";
+import EnterMeal from "../components/EnterMeal";
 
 export const getServerSideProps = optionalAuth;
